% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{summary.GpGp4RG_fit}
\alias{summary.GpGp4RG_fit}
\title{Print summary of GpGp4RG fit}
\usage{
\method{summary}{GpGp4RG_fit}(object, ...)
}
\arguments{
\item{object}{Object of class "GpGp4RG_fit", usually the return value from
\code{\link{fit_model}}}

\item{...}{additional arguments, for compatability with S3 generic 'summary'}
}
\description{
Print summary of GpGp4RG fit
}
